backend/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.5
pillow==11.0.0
PyJWT==2.10.0
PyPDF2==3.0.1
sqlparse==0.5.1
